serializer: Rename json/json.py to prot/jsonlib.py
It's so easy that conflicts with the standard module, json. It's
necessary to avoid this. Moreover, create a prot dir shorted for
"protocol" to keep available ones.
